backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
aiomysql
aiosqlite
pytest
httpx
dashscope
redis
python-dotenv
tenacity
jieba
pandas
scikit-learn
matplotlib
openpyxl


